feat(home-seeker): empty state dashboard
